refactor: split playground command handlers
