+++ packages/runtime-playground/src/wp-cli-command-handlers.ts
@@ -1,0 +1,70 @@
+import { argValue } from "./command-args.js"
+
+export function wpCliCommandFromArgs(args: string[]): string {
+  const explicit = argValue(args, "command")
+  if (explicit) {
+    return explicit.trim()
+  }
+
+  return args.join(" ").trim()
+}
+
+export function shellArgv(command: string): string[] {
+  const args: string[] = []
+  let current = ""
+  let quote = ""
+
+  for (let index = 0; index < command.length; index++) {
+    const char = command[index]
+    if (!quote && /\s/.test(char)) {
+      if (current) {
+        args.push(current)
+        current = ""
+      }
+      continue
+    }
+
+    if ((char === "'" || char === '"') && (!quote || quote === char)) {
+      quote = quote ? "" : char
+      continue
+    }
+
+    if (char === "\\" && index + 1 < command.length) {
+      current += command[++index]
+      continue
+    }
+
+    current += char
+  }
+
+  if (quote) {
+    throw new Error("Unclosed quote in wordpress.wp-cli command")
+  }
+
+  if (current) {
+    args.push(current)
+  }
+
+  return args
+}
+
+export function wpCliPhpScript(argv: string[]): string {
+  return `<?php
+putenv('SHELL_PIPE=0');
+$GLOBALS['argv'] = array_merge(array('/tmp/wp-cli.phar', '--path=/wordpress', '--no-color'), json_decode(${JSON.stringify(JSON.stringify(argv))}, true));
+if (!defined('STDIN')) {
+    define('STDIN', fopen('php://stdin', 'rb'));
+}
+if (!defined('STDOUT')) {
+    define('STDOUT', fopen('php://stdout', 'wb'));
+}
+if (!defined('STDERR')) {
+    define('STDERR', fopen('php://stderr', 'wb'));
+}
+require '/tmp/wp-cli.phar';
+`
+}
+
+export function cleanWpCliOutput(output: string): string {
+  return output.replace(/^#!\/usr\/bin\/env php\r?\n/, "")
+}
